AI_Agents/Phidata/requirements.txt
phidata
packaging
openai
yfinance
duckduckgo-search

